requirements.txt
ansible==6.6.0
ansible-core==2.13.6
appdirs==1.4.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
decorator==5.1.1
dogpile.cache==1.1.8
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.3
keystoneauth1==5.0.0
MarkupSafe==2.1.1
munch==2.5.0
netifaces==0.11.0
openstacksdk==0.61.0
os-service-types==1.7.0
packaging==21.3
pbr==5.11.0
pycparser==2.21
pyparsing==3.0.9
PyYAML==6.0
requests==2.28.1
requestsexceptions==1.4.0
resolvelib==0.8.1
six==1.16.0
stevedore==4.1.0
urllib3==1.26.12
